pyproject.toml
[project]
name = "healthcare-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "langchain-openai>=0.3.26",
    "langchain-text-splitters>=0.3.8",
    "pymupdf4llm>=0.0.25",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
